currency selector works for all price related data
