+++ src/app/components/MarketDataBar/MarketDataBar.tsx
@@ -5,7 +5,9 @@ import CoinCount from "./svg/CoinCount";
 import Btc from "./svg/Btc";
 import Eth from "./svg/Eth";
 import TotalVol from "./svg/TotalVol";
-import { MarketDataTypes } from "@/app/layout";
+import { useState, useEffect } from "react";
+import { useCoin } from "@/app/contexts/CoinProvider";
+import { CurrencyValue } from "types";
 
 const MarketDataBarWrapper = styled.div`
   background: #1e1932;
@@ -32,12 +34,72 @@ const SvgWrapper = styled.div`
   margin-right: 5px;
 `;
 
-interface MarketDataBarProps {
-  marketData: MarketDataTypes;
-  hasError: boolean;
+interface FetchedDataTypes {
+  data: {
+    total_market_cap: CurrencyValue;
+    total_volume: CurrencyValue;
+    market_cap_percentage: {
+      eth: number;
+      btc: number;
+    };
+    active_cryptocurrencies: number;
+  };
 }
 
-const MarketDataBar = ({ marketData, hasError }: MarketDataBarProps) => {
+export interface MarketDataTypes {
+  coins: string;
+  totalMarketCap: string;
+  totalVolume: string;
+  btcPercent: number;
+  ethPercent: number;
+}
+
+const MarketDataBar = () => {
+  const [marketData, setMarketData] = useState<MarketDataTypes | null>(null);
+  const [hasError, setHasError] = useState(false);
+
+  const { fiatCurrency } = useCoin();
+
+  function abbreviateNumber(num: number): string {
+    const prefixes = ["", "k", "M", "B", "T"];
+    const magnitude = Math.floor(Math.log10(Math.abs(num)) / 3);
+    const abbreviated = (num / Math.pow(1000, magnitude)).toFixed(2);
+    return `${abbreviated}${prefixes[magnitude]}`;
+  }
+
+  const apiKey = process.env.NEXT_PUBLIC_API_KEY;
+
+  useEffect(() => {
+    const getMarketData = async () => {
+      setHasError(false);
+      try {
+        const response = await fetch(
+          `https://pro-api.coingecko.com/api/v3/global?x_cg_pro_api_key=${apiKey}`
+        );
+        const fetchedData: FetchedDataTypes = await response.json();
+        const totalCap = abbreviateNumber(
+          fetchedData.data.total_market_cap[fiatCurrency]
+        );
+        const totalVol = abbreviateNumber(
+          fetchedData.data.total_volume[fiatCurrency]
+        );
+        const updatedMarketData = {
+          coins: Math.round(
+            fetchedData.data.active_cryptocurrencies
+          ).toLocaleString(),
+          btcPercent: Math.round(fetchedData.data.market_cap_percentage.btc),
+          ethPercent: Math.round(fetchedData.data.market_cap_percentage.eth),
+          totalMarketCap: totalCap,
+          totalVolume: totalVol,
+        };
+        setMarketData(updatedMarketData);
+      } catch {
+        setHasError(true);
+      }
+    };
+    getMarketData();
+  }, [fiatCurrency]);
+
   if (!marketData) {
     return <MarketDataBarWrapper>Loading...</MarketDataBarWrapper>;
   }
